refactor: 도메인 패키지 생성 & 도메인 클래스 이동
