+++ src/main/kotlin/lotto/domain/DefaultLottoMachine.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 object DefaultLottoMachine : LottoMachine {
