+++ src/main/kotlin/lotto/domain/LottoMachine.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 fun interface LottoMachine {
     fun generate(): LottoTicket
