+++ src/main/kotlin/lotto/domain/LottoNumber.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 data class LottoNumber(private val number: Int) {
