+++ src/main/kotlin/lotto/domain/LottoRanking.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 enum class LottoRanking(
     val matchCount: Int,
