+++ src/main/kotlin/lotto/domain/Money.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 import java.math.BigDecimal
 import java.math.RoundingMode
