+++ src/test/kotlin/lotto/domain/LottoShopTest.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.shouldBe
