+++ src/test/kotlin/lotto/domain/LottoTicketTest.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.DescribeSpec
